Rename cube.js to something coherent + maps index
