+++ index.html
@@ -10,7 +10,7 @@
 			<source src="audio.mp3" type="audio/ogg"> </audio>
 		<script src="js/three.js"></script>
 		<script src="js/tween.umd.js"></script>
-		<script type="module" src="js/cube.js"></script>
+		<script type="module" src="js/musical-labyrinth.js"></script>
 		<div id="container">
 			<div id="player"></div>
 			<div id="debugHover">
